Rename error -> handleError
